--- CapStone/react_frontend/AWARE/src/components/BillCard.jsx
@@ -1,9 +1,0 @@
-
-export const BillCard = ({ bill }) => {
-
-        return (
-            <div className="flex flex-col items-center justify-center text-center rounded-xl h-[25%] w-[98%] bg-slate-300">
-                {bill.title}
-            </div>
-        )
-}
