+++ CapStone/react_frontend/AWARE/src/components/Navbar.jsx
@@ -31,8 +31,8 @@ function Navbar() {
   }, []);
 
   return (
-    <div className='h-24 bg-zinc-800'>
-    <div className='navt flex justify-between items-center max-w-[80%] mx-auto px-4 pd-2 text-white bg-zinc-800'>
+    <div className='h-24 bg-[#001500]'>
+    <div className='navt flex justify-between items-center max-w-[80%] mx-auto px-4 pd-2 text-white bg-[#001500]'>
       <h1 className='w-full text-3xl font-bold text-white'>AWARE</h1>
       <ul className='navbar hidden md:flex '>
         <li className='navitem p-4'>
