+++ CapStone/react_frontend/AWARE/src/components/Politician.jsx
@@ -33,7 +33,7 @@ function borderColor(party) {
     else if (party == 'R'){
         return 'border-red-500'
     }
-    else { return('border-yellow-300') }
+    else { return('border-gray-600') }
 }
 
 
@@ -42,6 +42,8 @@ function borderColor(party) {
 function PoliticianBlock({ pol, image_url, isExpanded, toggleCollapse}) {
 
     const blockRef = useRef(null);
+
+
 
     useEffect(() => {
         if (isExpanded && blockRef.current) {
@@ -52,7 +54,7 @@ function PoliticianBlock({ pol, image_url, isExpanded, toggleCollapse}) {
     const cardClassCollapsed = `relative flex flex-col justify-center items-center bg-white gap-2 py-2 
                                 rounded-xl w-[45%] h-[2%] md:w-[30%] md:h-[3%] lg:w-[15%] lg:h-[7%]`
 
-    const cardClassExpanded = `relative flex-col justify-center items-center bg-white gap-2 py-2 
+    const cardClassExpanded = `relative flex flex-col justify-center items-center bg-white gap-2 py-2 
                                 rounded-xl w-[93%] h-[5%] md:w-[93%] md:h-[8%] lg:w-[94%] lg:h-[10%]`
 
     const imageClassCollapsed = `${borderColor(pol.party)} border-4 rounded-full overflow-hidden w-20 h-20 
@@ -81,7 +83,7 @@ function PoliticianBlock({ pol, image_url, isExpanded, toggleCollapse}) {
                 <div className="flex flex-col justify-center items-center">
                         <h5 className="card-title">{pol.first_name} {pol.last_name}</h5>
                         <p className="card-text">{pol.state}</p>
-                        {!isExpanded ? "" : <a href={pol.website} target="_blank">View Website</a>}
+                        {!isExpanded ? "" : <p>{pol.website} {pol.facebook}</p>}
                 </div>
             </div>
     );
